extract render_pdf to its own file and fallback to customer default address
