+++ test/render_pdf_test.rb
@@ -1,4 +1,5 @@
 require 'test_helper'
+require_relative '../app/utils/render_pdf'
 
 class RenderPdfTest < ActiveSupport::TestCase
 
@@ -20,6 +21,11 @@ class RenderPdfTest < ActiveSupport::TestCase
     pdf = render_pdf(@shop, order, @charity, 20)
   end
 
+  test "order_no_billing_address_uses_customer_default_address" do
+    order = JSON.parse(load_fixture('order_customer_address.json'))
+    pdf = render_pdf(@shop, order, @charity, 20)
+  end
+
   test "order_no_zip" do
     order = JSON.parse(load_fixture('order_no_zip.json'))
     pdf = render_pdf(@shop, order, @charity, 20)
@@ -32,22 +38,6 @@ class RenderPdfTest < ActiveSupport::TestCase
   end
 
   private
-
-  def render_pdf(shop, order, charity, donation_amount)
-    order['created_at'] = Time.parse(order['created_at']).strftime("%B %d, %Y")
-
-    template = Tilt::LiquidTemplate.new { |t| charity.pdf_template }
-    pdf_content = template.render(
-      shop: shop.attributes,
-      order: order,
-      charity: charity,
-      donation_amount: donation_amount
-    )
-
-    WickedPdf.new.pdf_from_string(
-      Tilt::ERBTemplate.new('views/receipt/pdf.erb').render(Object.new, pdf_content: pdf_content)
-    )
-  end
 
   def write_pdf(pdf_string)
     File.open('test.pdf', 'w') { |file| file.write(pdf_string) }
